timer: Change prescaler arugment to frequency argument in timer_init()
Change timer_init() prescaler argument, which says how the parent
clock frequnecy should be prescaled to result in the frequency we want,
to a frequency argument, which directly says which frequency we want.

The prescaler is then computed from the requested frequency and timer
parent frequency.

With this it will be simpler to have generic code for multiple
platforms.

Signed-off-by: Marek Behún <marek.behun@nic.cz>